Refactor Landing Page Components with Section Wrapper
